Fixe reading of decimal numbers when they are saved as int
The xls standard allows to save a decimal number like 1.75 as an
integer, multiplied by 100. The current implementation did neither
detect this as an integer nor as a float and subsequently returns the
default value 0.

This fix adds the case to RKNumber's Float64() method in order to
retrieve the value correctly as float.